Optional XI2.1 support
This patch adds XI2.1 support to mtview. Several instances of the the
program can be open simultaneously, each one being a window into the
device, which is mapped onto the whole screen. On a touchscreen,
paiting can occur simultaneously in each window. The device mapping is
automatically modified in multi-head setups, and xinput set-prop can
be used to map the device to a portion of the screen. Thanks to Chase
Douglas for fixing this.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>