Always build with XI 2.2 if it is available (except we don't)
Though we unconditionally enable XI 2.2 compilation now, we need to
force the XI 2 2 interface off until utouch-frame and/or mtview are fixed.
The frame-xi2.h header mtview requires is not installed and the following
symbols are not exported by the library:
    utouch_frame_configure_xi2
    utouch_frame_pump_xi2
    utouch_frame_is_supported_xi2
    utouch_frame_init_xi2

Until this is fixed, force the XI 2.2 interface off in the code.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>