Fix xi2 crash when ID limit is exceeded
After DIM_TOUCH touches have been registered by the program in
total, the program will crash when running the XI2 backend becuase
old IDs are never recycled. The 'detail' continues to rise and
rise with each new touch, eventually causing us to walk off the end
of the ID and colors arrays.

This tweaks our call to provide the index of the actual touch rather
than using the ID as an indirect index.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>