Get XI2 X and Y axes from valuators, not root
The root value is not in device space, which causes the re-mapping
done in the output function to no longer map up 1:1 with what is
on-screen. We can either pre-scale the root values or grab them
from the valuators to fix this. I go with the latter, being fairly
liberal in determining appropriate X and Y axes. If somehow we
find ourselves without an X or Y valuator in the middle of processing,
we exit.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>