Use sensible slot IDs in XI2 mode
After DIM_TOUCH touches have been registered by the program in
total, the program will crash when running the XI2 backend becuase
old IDs are never recycled. The 'next_slot' continues to rise with
each new touch, eventually causing the 'output_touch' function to
write byond the end of the 'id' and 'color' arrays.

This changes how slot IDs are generated in 'handle_xi2_event' so
that this doesn't occur. We use the index of the touch in the
'ti->touches' array as a de-facto slot identifier.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>