Access touch_data in handle_xi2_event through pointer, not index
The loop index 'i' is used in handle_xi2_event as both an index
to 'ti->touches' and as an index to the event's valuator mask.
Both senses are used simultaneously in the latter-half of the
function, leading to incorrect array accesses.

This function isolates the two usages by storing a pointer to
the touch_data found in the first half, so that the variable
may be safely reused in the second half.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>